requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==41.0.5
daphne==4.0.0
dj-rest-auth==5.0.2
Django==4.2.7
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
GDAL==3.8.0
geographiclib==2.0
geopy==2.4.1
haversine==2.8.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
msgpack==1.0.7
Pillow==10.1.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyotp==2.9.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.31.0
requests==2.31.0
rpds-py==0.13.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
urllib3==1.26.15
zope.interface==6.1
drf-extra-fields==3.7.0
firebase-admin==6.3.0
celery==5.3.6
qrcode==7.4.2
